Initialize django project
This is a standard django project configuration,
i have made some changes on the settings.py file
to accept some environment variables as asked on
the Trello board, since it increases the security
of the application. This is a standard quickstart
app with minor changes to the settings.py file.

The changes include:

-SECRET_KEY
-ALL DB VARS(HOST, PASSW,USER,PORT)
-DEBUG

Those were set on the settings.py file to get it
from the environment variables.